fix: rename logging file to logger
