+++ README.md
@@ -18,6 +18,5 @@
 
 중복선택 가능
 
-1. blob storage 이미지 올리기
 1. 스크린샷 검색 & 음성 추가
 1. 알람안받기 설정
